Fix warnings when building for msp430-elf
The MSP430 target supports both 16-bit and 20-bit size_t and intptr_t.
Some implicit casts in Newlib expect these types to be
"long", (a 32-bit type on MSP430) which causes warnings during
compilation such as:
  "cast from pointer to integer of different size"

+++ newlib/libc/stdio/open_memstream.c
@@ -326,8 +326,8 @@ internal_open_memstream_r (struct _reent *ptr,
   if (c->max < 64)
     c->max = 64;
 #if (SIZE_MAX >= 64 * 1024)
-  else if (c->max > 64 * 1024)
-    c->max = 64 * 1024;
+  else if (c->max > (size_t)64 * 1024)
+    c->max = (size_t)64 * 1024;
 #endif
   *size = 0;
   *buf = _malloc_r (ptr, c->max);
